Compute a bunch of statistics in ParserMain.

+++ src/main/scala/com/github/fedeoasi/model/DailyRideCount.scala
@@ -2,6 +2,6 @@ package com.github.fedeoasi.model
 
 import org.joda.time.LocalDate
 
-case class DailyRideCount(station: StationReference, date: LocalDate, dayType: String, rides: Int)
+case class DailyRideCount(station: StationReference, date: LocalDate, dayType: String, rides: Long)
 
 case class StationReference(stationId: Int, stationName: String)
